+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/cryptography_client_options.cpp
@@ -8,7 +8,5 @@ namespace Azure {
     namespace KeyVault {
       namespace Keys {
         namespace Cryptography {
-  const ServiceVersion ServiceVersion::V7_0("7.0");
-  const ServiceVersion ServiceVersion::V7_1("7.1");
   const ServiceVersion ServiceVersion::V7_2("7.2");
 }}}}} // namespace Azure::Security::KeyVault::Keys::Cryptography