+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/decrypt_parameters.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/keyvault/keys/cryptography/decrypt_parameters.hpp"
+#include <azure/core/base64.hpp>
+
 #include "../private/cryptography_serializers.hpp"
 #include "../private/key_constants.hpp"
-
-#include <azure/keyvault/common/internal/base64url.hpp>
+#include "azure/keyvault/keys/cryptography/decrypt_parameters.hpp"
 
 #include <string>
 #include <vector>
@@ -22,13 +22,14 @@ namespace Azure {
     Azure::Core::Json::_internal::json payload;
 
     using namespace Azure::Security::KeyVault::Keys::_detail;
-    using namespace Azure::Security::KeyVault::_internal;
+    using namespace Azure::Core::_internal;
     payload[AlgorithmValue] = parameters.Algorithm.ToString();
     payload[ValueParameterValue] = Base64Url::Base64UrlEncode(parameters.Ciphertext);
+    auto& iv = parameters.GetIv();
 
-    if (parameters.Iv.size() > 0)
+    if (iv.size() > 0)
     {
-      payload[IvValue] = Base64Url::Base64UrlEncode(parameters.Iv);
+      payload[IvValue] = Base64Url::Base64UrlEncode(iv);
     }
 
     if (parameters.AdditionalAuthenticatedData.size() > 0)