+++ sdk/keyvault/azure-security-keyvault-keys/src/key_client_options.cpp
@@ -4,7 +4,5 @@
 #include "azure/keyvault/keys/key_client_options.hpp"
 
 namespace Azure { namespace Security { namespace KeyVault { namespace Keys {
-  const ServiceVersion ServiceVersion::V7_0("7.0");
-  const ServiceVersion ServiceVersion::V7_1("7.1");
   const ServiceVersion ServiceVersion::V7_2("7.2");
 }}}} // namespace Azure::Security::KeyVault::Keys