+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -4,12 +4,19 @@
 
 ### Features Added
 
+- Added `GetIv()` to `EncryptParameters` and `DecryptParameters`.
+
 ### Breaking Changes
+
+- Removed `Azure::Security::KeyVault::Keys::ServiceVersion::V7_0` and `V7_1`.
+- Removed `Azure::Security::KeyVault::Keys::Cryptography::ServiceVersion::V7_0` and `V7_1`.
+- Removed `CryptographyClient::RemoteClient()` and `CryptographyClient::LocalOnly()`.
+- Removed the general constructor from `EncryptParameters` and `DecryptParameters`.
+- Removed access to `Iv` field member from `EncryptParameters` and `DecryptParameters`.
 
 ### Key Bugs Fixed
 
 ### Fixed
-
 
 ## 4.0.0-beta.3 (2021-06-08)